Support for parquet backend, annotation support, and dataframely support (#241)
## 0.10.0 (2025-07-04)
- Added ParquetTableStore which is based on duckdb SQLTableStore. It stores all tables as parquet files but still
  references them inside a duckdb database file as views to `FROM read_parquet(file)`. It supports both normal
  file systems and fsspec supported blob stores like AWS S3, Azure Blob Storage, and Google Cloud Storage.
- ColSpec / DataFramely support based on annotations of parameters and return type of tasks
- Pandas table hook reads date/datetime columns as 'datetime64[us]' and thus does not need clipping
  and extra year columns any more
- Materialize Local Table Cache before actual Table Store
- Mssql based SQLTableStore uses mssqlkit or bcpandas for bulk upload and arrow-odbc for download by default
- Postgres based SQLTableStore uses ADBC for download by default (already supports bulk upload)
- Fix numpy import issue on OS X
- CI: Use SQL Server docker image instead of Azure Edge
- Improve documentation for materialization details.
- Add support for columnstore tables in `MSSQLTableStore` via `MSSqlMaterializationDetails`

+++ src/pydiverse/pipedag/backend/blob.py
@@ -1,78 +1,24 @@
-from __future__ import annotations
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
 import os
 import pickle
 import shutil
-from abc import ABC, abstractmethod
-from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
+from upath import UPath
+
+from pydiverse.pipedag import Blob, Stage
 from pydiverse.pipedag.context import ConfigContext
 from pydiverse.pipedag.errors import CacheError
-from pydiverse.pipedag.util import Disposable, normalize_name
-
-if TYPE_CHECKING:
-    from pydiverse.pipedag.core import Stage
-    from pydiverse.pipedag.materialize import Blob
+from pydiverse.pipedag.materialize.store import BaseBlobStore
+from pydiverse.pipedag.util import normalize_name
 
 __all__ = [
-    "BaseBlobStore",
     "FileBlobStore",
 ]
 
-
-class BaseBlobStore(Disposable, ABC):
-    """Blob store base class
-
-    A blob (binary large object) store is responsible for storing arbitrary
-    python objects. This can, for example, be done by serializing them using
-    the python ``pickle`` module.
-
-    A store must use a blob's name (``Blob.name``) and stage (``Blob.stage``)
-    as the primary keys for storing and retrieving blobs. This means that
-    two different ``Blob`` objects can be used to store and retrieve the same
-    data as long as they have the same name and stage.
-    """
-
-    @abstractmethod
-    def init_stage(self, stage: Stage):
-        """Initialize a stage and start a transaction"""
-
-    @abstractmethod
-    def commit_stage(self, stage: Stage):
-        """Commit the stage transaction
-
-        Replace the blobs of the base stage with the blobs in the transaction.
-        """
-
-    @abstractmethod
-    def store_blob(self, blob: Blob):
-        """Stores a blob in the associated stage transaction"""
-
-    @abstractmethod
-    def copy_blob_to_transaction(self, blob: Blob):
-        """Copy a blob from the base stage to the transaction
-
-        This operation MUST not remove the blob from the base stage or modify
-        it in any way.
-        """
-
-    @abstractmethod
-    def delete_blob_from_transaction(self, blob: Blob):
-        """Delete a blob from the transaction
-
-        If the blob doesn't exist in the transaction, fail silently.
-        """
-
-    @abstractmethod
-    def retrieve_blob(self, blob: Blob) -> Any:
-        """Loads a blob from the store
-
-        Retrieves the stored python object from the store and returns it.
-        If the stage hasn't yet been committed, the blob must be retrieved
-        from the transaction, else it must be retrieved from the committed
-        stage.
-        """
+from pydiverse.pipedag.util.path import is_file_uri
 
 
 class NoBlobStore(BaseBlobStore):
@@ -128,61 +74,73 @@ class FileBlobStore(BaseBlobStore):
     @classmethod
     def _init_conf_(cls, config: dict[str, Any]):
         instance_id = normalize_name(ConfigContext.get().instance_id)
-        base_path = Path(config["base_path"]) / instance_id
+        base_path = UPath(config["base_path"]) / instance_id
         return cls(base_path)
 
-    def __init__(self, base_path: str | Path):
-        self.base_path = Path(base_path).absolute()
-        os.makedirs(self.base_path, exist_ok=True)
+    def __init__(self, base_path: str | UPath):
+        self.base_path = UPath(base_path).absolute()
+        if is_file_uri(self.base_path):
+            os.makedirs(self.base_path, exist_ok=True)
 
     def init_stage(self, stage: Stage):
         stage_path = self.get_stage_path(stage.name)
         transaction_path = self.get_stage_path(stage.transaction_name)
 
-        try:
-            os.mkdir(stage_path)
-        except FileExistsError:
-            pass
+        if is_file_uri(stage_path):
+            try:
+                os.mkdir(stage_path)
+            except FileExistsError:
+                pass
 
-        try:
-            os.mkdir(transaction_path)
-        except FileExistsError:
-            shutil.rmtree(transaction_path)
-            os.mkdir(transaction_path)
+        if is_file_uri(transaction_path):
+            try:
+                os.mkdir(transaction_path)
+            except FileExistsError:
+                shutil.rmtree(transaction_path)
+                os.mkdir(transaction_path)
+        else:
+            # clear bucket key prefix similar to removing a directory
+            if transaction_path.exists():
+                transaction_path.fs.rmdir(transaction_path)
 
     def commit_stage(self, stage: Stage):
         stage_path = self.get_stage_path(stage.name)
         transaction_path = self.get_stage_path(stage.transaction_name)
         tmp_path = self.get_stage_path(stage.name + "__swap")
 
+        if not is_file_uri(transaction_path):
+            # TODO: think about switching to odd/even transaction schemas
+            # to avoid renaming all files in a schema.
+            # Another alternative: store all files written to transaction
+            raise NotImplementedError("Using Bucket as base path is currently not implemented.")
         os.rename(transaction_path, tmp_path)
         os.rename(stage_path, transaction_path)
         os.rename(tmp_path, stage_path)
         shutil.rmtree(transaction_path)
 
     def store_blob(self, blob: Blob):
-        with open(
-            self.get_blob_path(blob.stage.transaction_name, blob.name), "wb"
-        ) as f:
+        with open(self.get_blob_path(blob.stage.transaction_name, blob.name), "wb") as f:
             pickle.dump(blob.obj, f, pickle.HIGHEST_PROTOCOL)
 
     def copy_blob_to_transaction(self, blob: Blob):
-        try:
-            shutil.copy2(
-                self.get_blob_path(blob.stage.name, blob.name),
-                self.get_blob_path(blob.stage.transaction_name, blob.name),
-            )
-        except FileNotFoundError:
-            raise CacheError(
-                f"Can't copy blob '{blob.name}' (stage: '{blob.stage.name}')"
-                " to working transaction because no such blob exists."
-            ) from None
+        from_path = self.get_blob_path(blob.stage.name, blob.name)
+        to_path = self.get_blob_path(blob.stage.transaction_name, blob.name)
+        if is_file_uri(from_path):
+            try:
+                shutil.copy2(
+                    from_path,
+                    to_path,
+                )
+            except FileNotFoundError:
+                raise CacheError(
+                    f"Can't copy blob '{blob.name}' (stage: '{blob.stage.name}')"
+                    " to working transaction because no such blob exists."
+                ) from None
+        else:
+            from_path.fs.copy(from_path, to_path, recursive=True)
 
     def delete_blob_from_transaction(self, blob: Blob):
-        try:
-            os.remove(self.get_blob_path(blob.stage.transaction_name, blob.name))
-        except FileNotFoundError:
-            return
+        self.get_blob_path(blob.stage.transaction_name, blob.name).unlink(missing_ok=True)
 
     def retrieve_blob(self, blob: Blob):
         stage = blob.stage
@@ -190,8 +148,8 @@ class FileBlobStore(BaseBlobStore):
         with open(self.get_blob_path(stage.current_name, blob.name), "rb") as f:
             return pickle.load(f)
 
-    def get_stage_path(self, stage_name: str) -> Path:
+    def get_stage_path(self, stage_name: str) -> UPath:
         return self.base_path / stage_name
 
-    def get_blob_path(self, stage_name: str, blob_name: str) -> Path:
+    def get_blob_path(self, stage_name: str, blob_name: str) -> UPath:
         return self.base_path / stage_name / (blob_name + ".pkl")
